requirements.txt
pandas
numpy
scikit-learn
scipy
google-cloud-storage